+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN DT VERSION 0.24
 
+- Fixed a bug with missing numeric values rendered as 0 (@tverbeke #1000).
+
 
 # CHANGES IN DT VERSION 0.23