+++ NEWS.md
@@ -1,3 +1,6 @@
+# CHANGES IN DT VERSION 0.25
+
+
 # CHANGES IN DT VERSION 0.24
 
 - Fixed a bug with missing numeric values rendered as `0` (thanks, @tverbeke, #1000).